+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "python.pythonPath": "/Users/francescocolonnese/miniconda3/envs/deepucla/bin/python"
+}
